chore: split ConfigurationService
