+++ src/central-server/admin-service/core/src/main/java/org/niis/xroad/cs/admin/core/service/ConfigurationServiceImpl.java
@@ -27,8 +27,6 @@
 package org.niis.xroad.cs.admin.core.service;
 
 import ee.ria.xroad.common.SystemProperties;
-import ee.ria.xroad.common.conf.globalconf.privateparameters.v2.ConfigurationAnchorType;
-import ee.ria.xroad.common.conf.globalconf.privateparameters.v2.ObjectFactory;
 import ee.ria.xroad.common.util.CryptoUtils;
 import ee.ria.xroad.common.util.process.ExternalProcessRunner;
 import ee.ria.xroad.commonui.OptionalConfPart;
@@ -37,51 +35,31 @@ import ee.ria.xroad.commonui.OptionalPartsConf;
 import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
 import lombok.SneakyThrows;
-import lombok.extern.slf4j.Slf4j;
-import org.apache.commons.lang3.StringUtils;
 import org.niis.xroad.cs.admin.api.domain.ConfigurationSourceType;
 import org.niis.xroad.cs.admin.api.domain.DistributedFile;
-import org.niis.xroad.cs.admin.api.dto.ConfigurationAnchor;
 import org.niis.xroad.cs.admin.api.dto.ConfigurationParts;
 import org.niis.xroad.cs.admin.api.dto.File;
 import org.niis.xroad.cs.admin.api.dto.GlobalConfDownloadUrl;
 import org.niis.xroad.cs.admin.api.dto.HAConfigStatus;
 import org.niis.xroad.cs.admin.api.exception.ConfigurationPartException;
-import org.niis.xroad.cs.admin.api.exception.ConfigurationSourceException;
 import org.niis.xroad.cs.admin.api.exception.NotFoundException;
 import org.niis.xroad.cs.admin.api.service.ConfigurationService;
 import org.niis.xroad.cs.admin.api.service.SystemParameterService;
-import org.niis.xroad.cs.admin.core.entity.ConfigurationSigningKeyEntity;
 import org.niis.xroad.cs.admin.core.entity.ConfigurationSourceEntity;
 import org.niis.xroad.cs.admin.core.entity.DistributedFileEntity;
 import org.niis.xroad.cs.admin.core.entity.mapper.DistributedFileMapper;
 import org.niis.xroad.cs.admin.core.repository.ConfigurationSourceRepository;
 import org.niis.xroad.cs.admin.core.repository.DistributedFileRepository;
 import org.niis.xroad.restapi.config.audit.AuditDataHelper;
-import org.niis.xroad.restapi.config.audit.AuditEventHelper;
 import org.springframework.stereotype.Service;
-import org.springframework.util.CollectionUtils;
 
 import javax.transaction.Transactional;
-import javax.xml.bind.JAXBContext;
-import javax.xml.bind.JAXBElement;
-import javax.xml.bind.JAXBException;
-import javax.xml.bind.Marshaller;
-import javax.xml.datatype.DatatypeConfigurationException;
-import javax.xml.datatype.DatatypeFactory;
 
-import java.io.StringWriter;
-import java.io.Writer;
-import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.time.Instant;
-import java.time.ZoneId;
-import java.time.ZonedDateTime;
-import java.util.GregorianCalendar;
 import java.util.HashSet;
 import java.util.List;
-import java.util.Optional;
 import java.util.Set;
 
 import static ee.ria.xroad.common.conf.globalconf.ConfigurationConstants.CONTENT_ID_PRIVATE_PARAMETERS;
@@ -92,15 +70,11 @@ import static ee.ria.xroad.common.util.CryptoUtils.DEFAULT_UPLOAD_FILE_HASH_ALGO
 import static java.util.stream.Collectors.toSet;
 import static org.apache.commons.lang3.StringUtils.isNotBlank;
 import static org.niis.xroad.cs.admin.api.domain.ConfigurationSourceType.EXTERNAL;
+import static org.niis.xroad.cs.admin.api.domain.ConfigurationSourceType.INTERNAL;
 import static org.niis.xroad.cs.admin.api.exception.ErrorMessage.CONFIGURATION_NOT_FOUND;
 import static org.niis.xroad.cs.admin.api.exception.ErrorMessage.CONFIGURATION_PART_FILE_NOT_FOUND;
 import static org.niis.xroad.cs.admin.api.exception.ErrorMessage.CONFIGURATION_PART_VALIDATION_FAILED;
-import static org.niis.xroad.cs.admin.api.exception.ErrorMessage.ERROR_RECREATING_ANCHOR;
-import static org.niis.xroad.cs.admin.api.exception.ErrorMessage.INSTANCE_IDENTIFIER_NOT_SET;
-import static org.niis.xroad.cs.admin.api.exception.ErrorMessage.NO_CONFIGURATION_SIGNING_KEYS_CONFIGURED;
 import static org.niis.xroad.cs.admin.api.exception.ErrorMessage.UNKNOWN_CONFIGURATION_PART;
-import static org.niis.xroad.restapi.config.audit.RestApiAuditEvent.RE_CREATE_EXTERNAL_CONFIGURATION_ANCHOR;
-import static org.niis.xroad.restapi.config.audit.RestApiAuditEvent.RE_CREATE_INTERNAL_CONFIGURATION_ANCHOR;
 import static org.niis.xroad.restapi.config.audit.RestApiAuditProperty.CONTENT_IDENTIFIER;
 import static org.niis.xroad.restapi.config.audit.RestApiAuditProperty.PART_FILE_NAME;
 import static org.niis.xroad.restapi.config.audit.RestApiAuditProperty.SOURCE_TYPE;
@@ -108,22 +82,11 @@ import static org.niis.xroad.restapi.config.audit.RestApiAuditProperty.UPLOAD_FI
 import static org.niis.xroad.restapi.config.audit.RestApiAuditProperty.UPLOAD_FILE_HASH_ALGORITHM;
 import static org.niis.xroad.restapi.config.audit.RestApiAuditProperty.UPLOAD_FILE_NAME;
 
-@Slf4j
 @Service
 @Transactional
 @RequiredArgsConstructor
 public class ConfigurationServiceImpl implements ConfigurationService {
-    private static final JAXBContext JAXB_CTX;
 
-    static {
-        try {
-            JAXB_CTX = JAXBContext.newInstance(ObjectFactory.class);
-        } catch (JAXBException e) {
-            throw new RuntimeException("Failed to initialize JAXB context", e);
-        }
-    }
-
-    private static final String INTERNAL_CONFIGURATION = "INTERNAL";
     private static final Set<String> NODE_LOCAL_CONTENT_IDS = Set.of(
             CONTENT_ID_PRIVATE_PARAMETERS,
             CONTENT_ID_SHARED_PARAMETERS);
@@ -133,14 +96,12 @@ public class ConfigurationServiceImpl implements ConfigurationService {
     private final ConfigurationSourceRepository configurationSourceRepository;
     private final DistributedFileRepository distributedFileRepository;
     private final DistributedFileMapper distributedFileMapper;
-    private final AuditEventHelper auditEventHelper;
     private final AuditDataHelper auditDataHelper;
     private final ExternalProcessRunner externalProcessRunner;
 
     @Override
     public Set<ConfigurationParts> getConfigurationParts(ConfigurationSourceType sourceType) {
-        final ConfigurationSourceEntity configurationSource = findConfigurationSourceBySourceType(
-                sourceType.name().toLowerCase());
+        final ConfigurationSourceEntity configurationSource = findConfigurationSourceBySourceType(sourceType);
         final String haNodeName = configurationSource.getHaNodeName();
 
         Set<ConfigurationParts> configurationParts = new HashSet<>();
@@ -219,92 +180,9 @@ public class ConfigurationServiceImpl implements ConfigurationService {
     }
 
     @Override
-    public ConfigurationAnchor getConfigurationAnchor(String sourceType) {
-        final ConfigurationSourceEntity configurationSource = findConfigurationSourceBySourceType(
-                sourceType.toLowerCase());
-
-        return new ConfigurationAnchor(configurationSource.getAnchorFileHash(),
-                configurationSource.getAnchorGeneratedAt());
-    }
-
-    @Override
-    public ConfigurationAnchor getConfigurationAnchorWithFile(String sourceType) {
-        final ConfigurationSourceEntity configurationSource = findConfigurationSourceBySourceType(
-                sourceType.toLowerCase());
-
-        return new ConfigurationAnchor(configurationSource.getAnchorFile(),
-                configurationSource.getAnchorFileHash(),
-                configurationSource.getAnchorGeneratedAt());
-    }
-
-    @Override
-    public ConfigurationAnchor recreateAnchor(String configurationType) {
-        auditEventHelper.changeRequestScopedEvent(configurationType.equals(INTERNAL_CONFIGURATION)
-                ? RE_CREATE_INTERNAL_CONFIGURATION_ANCHOR
-                : RE_CREATE_EXTERNAL_CONFIGURATION_ANCHOR);
-
-        final var instanceIdentifier = Optional.ofNullable(systemParameterService.getInstanceIdentifier())
-                .filter(StringUtils::isNotEmpty)
-                .orElseThrow(() -> new ConfigurationSourceException(INSTANCE_IDENTIFIER_NOT_SET));
-
-        final var configurationSource = configurationSourceRepository.findBySourceTypeOrCreate(
-                configurationType.toLowerCase(),
-                haConfigStatus);
-
-        if (CollectionUtils.isEmpty(configurationSource.getConfigurationSigningKeys())) {
-            throw new ConfigurationSourceException(NO_CONFIGURATION_SIGNING_KEYS_CONFIGURED);
-        }
-
-        final var sources = configurationSourceRepository.findAllBySourceType(configurationType.toLowerCase());
-        final var now = ZonedDateTime.now(ZoneId.of("UTC"));
-        final var anchorXml = buildAnchorXml(configurationType, instanceIdentifier, now, sources);
-        final var anchorXmlBytes = anchorXml.getBytes(StandardCharsets.UTF_8);
-        final var anchorXmlHash = auditDataHelper.putAnchorHash(anchorXmlBytes);
-        for (final var src : sources) {
-            if (src.getConfigurationSigningKey() != null) {
-                src.setAnchorGeneratedAt(now.toInstant());
-                src.setAnchorFileHash(anchorXmlHash);
-                src.setAnchorFile(anchorXmlBytes);
-                configurationSourceRepository.save(src);
-            }
-        }
-
-        return new ConfigurationAnchor(anchorXmlHash, now.toInstant());
-    }
-
-    private String buildAnchorXml(final String configurationType,
-                                  final String instanceIdentifier,
-                                  final ZonedDateTime now,
-                                  final List<ConfigurationSourceEntity> sources) {
-        try {
-
-            Marshaller marshaller = JAXB_CTX.createMarshaller();
-            marshaller.setProperty(Marshaller.JAXB_FORMATTED_OUTPUT, Boolean.TRUE);
-
-
-            final var factory = new ObjectFactory();
-            final var configurationAnchor = factory.createConfigurationAnchorType();
-            configurationAnchor.setGeneratedAt(DatatypeFactory.newInstance().newXMLGregorianCalendar(GregorianCalendar.from(now)));
-            configurationAnchor.setInstanceIdentifier(instanceIdentifier);
-
-            sources.stream()
-                    .map(src -> toXmlSource(src, configurationType, factory))
-                    .forEach(configurationAnchor.getSource()::add);
-
-            JAXBElement<ConfigurationAnchorType> root = factory.createConfigurationAnchor(configurationAnchor);
-
-            Writer writer = new StringWriter();
-            marshaller.marshal(root, writer);
-            return writer.toString();
-        } catch (DatatypeConfigurationException | JAXBException e) {
-            throw new ConfigurationSourceException(ERROR_RECREATING_ANCHOR);
-        }
-    }
-
-    @Override
-    public GlobalConfDownloadUrl getGlobalDownloadUrl(String sourceType) {
+    public GlobalConfDownloadUrl getGlobalDownloadUrl(ConfigurationSourceType sourceType) {
         final String csAddress = systemParameterService.getCentralServerAddress();
-        final String sourceDirectory = sourceType.equals(INTERNAL_CONFIGURATION)
+        final String sourceDirectory = sourceType.equals(INTERNAL)
                 ? SystemProperties.getCenterInternalDirectory()
                 : SystemProperties.getCenterExternalDirectory();
 
@@ -330,7 +208,6 @@ public class ConfigurationServiceImpl implements ConfigurationService {
                 .map(distributedFileMapper::toTarget)
                 .collect(toSet());
     }
-
 
     @Override
     public void uploadConfigurationPart(ConfigurationSourceType sourceType,
@@ -398,35 +275,10 @@ public class ConfigurationServiceImpl implements ConfigurationService {
         return NODE_LOCAL_CONTENT_IDS.contains(contentId);
     }
 
-    private ConfigurationSourceEntity findConfigurationSourceBySourceType(String sourceType) {
-        return configurationSourceRepository.findBySourceTypeAndHaNodeName(sourceType, haConfigStatus.getCurrentHaNodeName())
-                .orElseThrow(ConfigurationServiceImpl::notFoundException);
+    private ConfigurationSourceEntity findConfigurationSourceBySourceType(ConfigurationSourceType sourceType) {
+        return configurationSourceRepository.findBySourceTypeAndHaNodeName(sourceType.name().toLowerCase(),
+                        haConfigStatus.getCurrentHaNodeName())
+                .orElseThrow(() -> new NotFoundException(CONFIGURATION_NOT_FOUND));
     }
 
-    private String buildGlobalDownloadUrl(final String sourceType, final String haNodeName) {
-        final var csAddress = systemParameterService.getCentralServerAddress(haNodeName);
-        final String sourceDirectory = sourceType.equals(INTERNAL_CONFIGURATION)
-                ? SystemProperties.getCenterInternalDirectory()
-                : SystemProperties.getCenterExternalDirectory();
-
-        return String.format("http://%s/%s", csAddress, sourceDirectory);
-    }
-
-    private ee.ria.xroad.common.conf.globalconf.privateparameters.v2.ConfigurationSourceType toXmlSource(
-            final ConfigurationSourceEntity source,
-            final String configurationType,
-            final ObjectFactory factory) {
-        final var xmlSource = factory.createConfigurationSourceType();
-
-        xmlSource.setDownloadURL(buildGlobalDownloadUrl(configurationType, source.getHaNodeName()));
-        source.getConfigurationSigningKeys().stream()
-                .map(ConfigurationSigningKeyEntity::getCert)
-                .forEach(xmlSource.getVerificationCert()::add);
-
-        return xmlSource;
-    }
-
-    private static NotFoundException notFoundException() {
-        return new NotFoundException(CONFIGURATION_NOT_FOUND);
-    }
 }
